# services/sales-rest/src/controllers/user/openAPI/openAPIDocumentation.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Sales Middleware API",
        "description": "Sales Middleware use case based, using SAP Sales Cloud"
    },
    "servers": [
        {
            "url": "https://sales-middleware-rest.c-416e85f.kyma.shoot.live.k8s-hana.ondemand.com/api/v1/sales",
            "description": "Kyma Middleware DEV API server"
        }
    ],
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "in": "headers",
            "name": "authorization"
        }
    },
    "paths": {
...
        "/tasks": {
            "get": {
                "summary": "GET Tasks",
                "tags": [
                    "Tasks"
                ],
                "description": "Get All tasks owned by the given UserId",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "header",
                        "required": false,
                        "description": "Current UserId",
                        "default": "00163E8B-259B-1EDA-9E98-432D21F5D55E"
                    },
                    {
                        "name": "status",
                        "description": "Task Status",
                        "in": "query",
                        "default": "OPEN",
                        "required": false,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    {
                        "name": "account",
                        "description": "Task account Id",
                        "in": "query",
                        "default": "00163E08ABF01ED69CE6DD15F43262CD",
                        "required": false,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "summary": "Create new task",
                "tags": [
                    "Tasks"
                ],
                "description": "Create new task",
                "parameters": [],
                "requestBody": {
                    "description": "new task payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "required": false,
                                        "example": "standalone task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "required": false,
                                        "example": "create new task from Open API"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "required": false,
                                        "example": "2020-06-24T09:30:00.000Z"
                                    },
                                    "highPriority": {
                                        "type": "boolean",
                                        "required": false,
                                        "example": false
                                    },
                                    "contact": {
                                        "type": "string",
                                        "required": false,
                                        "example": "00163E08ABF01EE78DC2DD0503029384"
                                    },
                                    "account": {
                                        "type": "string",
                                        "required": false,
                                        "example": "00163E08ABF01EE78DC2918D620D5224"
                                    },
                                    "processor": {
                                        "type": "string",
                                        "required": false,
                                        "example": "00163E08ABF01EE69CEDCBE3E05D1273"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Successfully  created  standalone task",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "404": {
                        "description": "BAD REQUEST"
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "summary": "GET Task",
                "tags": [
                    "Tasks"
                ],
                "description": "Get Task by ID",
                "parameters": [
                    {
                        "name": "taskId",
                        "description": "Task ID",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "default": "00163E7930751EDAB68084B352C5B89E"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Task": {
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "updateDate": {
                    "type": "string"
                },
                "issue": {
                    "type": "object"
                },
                "primaryContact": {
                    "type": "object"
                }
            }
        },
...
    }
}